# app/Models/User.php
<?php

namespace App\Models;

use Tymon\JWTAuth\Contracts\JWTSubject;
use Illuminate\Contracts\Auth\MustVerifyEmail;
use Illuminate\Foundation\Auth\User as Authenticatable;
use Illuminate\Notifications\Notifiable;

/**
 * Class User
 *
 */
/**
 * @SWG\Definition()
 */
class User extends Authenticatable implements JWTSubject
{
    use Notifiable;

    /**
     * The attributes that are mass assignable.
     *
     * @var array
     */
    protected $fillable = [
        'name', 'email', 'password',
    ];

    /**
     * The attributes that should be hidden for arrays.
     *
     * @var array
     */
    protected $hidden = [
        'password', 'remember_token',
    ];

    /**
     * The attributes that should be cast to native types.
     *
     * @var array
     */
    protected $casts = [
        'email_verified_at' => 'datetime',
    ];

    /**
     * Get the identifier that will be stored in the subject claim of the JWT.
     *
     * @return mixed
     */
    public function token()
    {
        return $this->getKey();
    }

    /**
     * Get the identifier that will be stored in the subject claim of the JWT.
     *
     * @return mixed
     */
    public function getJWTIdentifier()
    {
        return $this->getKey();
    }

    /**
     * Return a key value array, containing any custom claims to be added to the JWT.
     *
     * @return array
     */
    public function getJWTCustomClaims()
    {
        return [];
    }

    /**
     *
     * @var number
     *
     * @SWG\Property(
     *   property="id",
     *   type="number"
     * )
     */
    /**
     *
     * @var string
     *
     * @SWG\Property(
     *   property="name",
     *   type="string"
     * )
     */
    /**
     *
     * @var string
     *
     * @SWG\Property(
     *   property="email",
     *   type="string"
     * )
     */
    /**
     *
     * @var string
     *
     * @SWG\Property(
     *   property="password",
     *   type="string"
     * )
     */
    /**
     * @var string
     *
     * @SWG\Property(
     *   property="remember_token",
     *   type="string"
     * )
     */
    /**
     * @var string
     *
     * @SWG\Property(
     *   property="created_at",
     *   type="string"
     * )
     */
    /**
     * @var string
     *
     * @SWG\Property(
     *   property="updated_at",
     *   type="string"
     * )
     */
}